requirements.txt
dockerfile-parse
gitdb
GitPython
six
requests